Adds status views
