+++ bookwyrm/views/__init__.py
@@ -7,3 +7,4 @@ from .notifications import Notifications
 from .direct_message import DirectMessage
 from .import_data import Import, ImportStatus
 from .user import User, EditUser, Followers, Following
+from .status import Status, Replies, CreateStatus, DeleteStatus
